Refactor: 프로젝트 구조 변경
- 멀티 프로젝트로 개선 하고자 프로젝트 구조변경함
