+++ collector/src/main/java/com/javacafe/realtime_sujeong/collector/CollectorApplication.java
@@ -1,0 +1,13 @@
+package com.javacafe.realtime_sujeong.collector;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"com.javacafe.realtime_sujeong.collector", "com.javacafe.realtime_sujeong.common"})
+public class CollectorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CollectorApplication.class, args);
+	}
+
+}
