+++ processor/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    implementation project(':common')
+    implementation 'org.springframework.kafka:spring-kafka'
+    
+    testImplementation 'org.springframework.kafka:spring-kafka-test'
+    testImplementation 'org.testcontainers:kafka'
+}
