+++ processor/src/main/java/com/javacafe/realtime_sujeong/processor/ProcessorApplication.java
@@ -1,0 +1,13 @@
+package com.javacafe.realtime_sujeong.processor;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"com.javacafe.realtime_sujeong.processor", "com.javacafe.realtime_sujeong.common"})
+public class ProcessorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProcessorApplication.class, args);
+	}
+
+}
