+++ serving/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    implementation project(':common')
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    runtimeOnly 'com.mysql:mysql-connector-j'
+    
+    testImplementation 'com.navercorp.fixturemonkey:fixture-monkey-starter:0.6.5'
+}
