+++ serving/src/main/java/com/javacafe/realtime_sujeong/serving/user/dto/UserCreateRequest.java
@@ -1,0 +1,39 @@
+package com.javacafe.realtime_sujeong.serving.user.dto;
+
+import com.javacafe.realtime_sujeong.serving.user.entity.User;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class UserCreateRequest {
+
+    @NotBlank(message = "이메일은 필수입니다.")
+    @Email(message = "올바른 이메일 형식이 아닙니다.")
+    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$", message = "올바른 이메일 형식이 아닙니다.")
+    private String email;
+
+    @NotBlank(message = "이름은 필수입니다.")
+    private String name;
+
+    @NotBlank(message = "비밀번호는 필수입니다.")
+    @Pattern(
+            regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()_+=-]).{8,}$",
+            message = "비밀번호는 8자 이상이며, 영문자, 숫자, 특수문자를 포함해야 합니다."
+    )
+    private String password;
+
+    public User toEntity() {
+        return User.builder()
+                .email(this.email)
+                .name(this.name)
+                .password(this.password)
+                .build();
+    }
+}
