+++ serving/src/main/java/com/javacafe/realtime_sujeong/serving/user/dto/UserUpdateRequest.java
@@ -1,0 +1,28 @@
+package com.javacafe.realtime_sujeong.serving.user.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.Getter;
+
+@Getter
+public class UserUpdateRequest {
+
+    @NotBlank(message = "id는 필수입니다.")
+    private long id;
+
+    @NotBlank(message = "이메일은 필수입니다.")
+    @Email(message = "올바른 이메일 형식이 아닙니다.")
+    private String email;
+
+    @NotBlank(message = "이름은 필수입니다.")
+    private String name;
+
+    @NotBlank(message = "비밀번호는 필수입니다.")
+    @Pattern(
+            regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()_+=-]).{8,}$",
+            message = "비밀번호는 8자 이상이며, 영문자, 숫자, 특수문자를 포함해야 합니다."
+    )
+    private String password;
+
+}
