+++ serving/src/main/java/com/javacafe/realtime_sujeong/serving/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.javacafe.realtime_sujeong.serving.user.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.javacafe.realtime_sujeong.serving.user.entity.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+}
