+++ settings.gradle
@@ -1,1 +1,7 @@
 rootProject.name = 'realtime-sujeong'
+
+include 'common'
+include 'collector' 
+include 'processor'
+include 'indexer'
+include 'serving'
