--- src/main/java/com/javacafe/realtime_sujeong/RealtimeSujeongApplication.java
@@ -1,15 +1,0 @@
-package com.javacafe.realtime_sujeong;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class RealtimeSujeongApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RealtimeSujeongApplication.class, args);
-	}
-
-}
