--- src/main/java/com/javacafe/realtime_sujeong/common/BaseEntity.java
@@ -1,24 +1,0 @@
-package com.javacafe.realtime_sujeong.common;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public class BaseEntity {
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column
-    LocalDateTime updatedAt;
-}
