--- src/main/java/com/javacafe/realtime_sujeong/user/controller/UserController.java
@@ -1,51 +1,0 @@
-package com.javacafe.realtime_sujeong.user.controller;
-
-import java.util.List;
-
-import jakarta.validation.Valid;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import com.javacafe.realtime_sujeong.user.dto.UserCreateRequest;
-import com.javacafe.realtime_sujeong.user.dto.UserResponse;
-import com.javacafe.realtime_sujeong.user.dto.UserUpdateRequest;
-import com.javacafe.realtime_sujeong.user.service.UserService;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/users")
-public class UserController {
-
-    private final UserService userService;
-
-    @GetMapping
-    public ResponseEntity<List<UserResponse>> findAllUsers() {
-        return ResponseEntity.ok(userService.findAllUsers());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<UserResponse> findUserById(@PathVariable long id) {
-        return ResponseEntity.ok(userService.findUserById(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<Void> createUser(@RequestBody @Valid UserCreateRequest userCreateRequest) {
-        userService.createUser(userCreateRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping
-    public ResponseEntity<Void> updateUser(@RequestBody UserUpdateRequest userUpdateRequest) {
-        userService.updateUser(userUpdateRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable long id) {
-        userService.deleteUserById(id);
-        return ResponseEntity.ok().build();
-    }
-
-}
