--- src/main/java/com/javacafe/realtime_sujeong/user/entity/User.java
@@ -1,41 +1,0 @@
-package com.javacafe.realtime_sujeong.user.entity;
-
-import com.javacafe.realtime_sujeong.common.BaseEntity;
-import com.javacafe.realtime_sujeong.user.dto.UserUpdateRequest;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-@AllArgsConstructor
-public class User extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String name;
-
-    public void update(UserUpdateRequest userUpdateRequest) {
-        this.name = userUpdateRequest.getName();
-        this.password = userUpdateRequest.getPassword();
-    }
-}
