--- src/test/java/com/javacafe/realtime_sujeong/RealtimeSujeongApplicationTests.java
@@ -1,15 +1,0 @@
-package com.javacafe.realtime_sujeong;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-
-@Import(TestcontainersConfiguration.class)
-@SpringBootTest
-class RealtimeSujeongApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
