--- src/test/java/com/javacafe/realtime_sujeong/TestcontainersConfiguration.java
@@ -1,25 +1,0 @@
-package com.javacafe.realtime_sujeong;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
-import org.testcontainers.elasticsearch.ElasticsearchContainer;
-import org.testcontainers.kafka.KafkaContainer;
-import org.testcontainers.utility.DockerImageName;
-
-@TestConfiguration(proxyBeanMethods = false)
-class TestcontainersConfiguration {
-
-	@Bean
-	@ServiceConnection
-	ElasticsearchContainer elasticsearchContainer() {
-		return new ElasticsearchContainer(DockerImageName.parse("docker.elastic.co/elasticsearch/elasticsearch:7.17.10"));
-	}
-
-	@Bean
-	@ServiceConnection
-	KafkaContainer kafkaContainer() {
-		return new KafkaContainer(DockerImageName.parse("apache/kafka-native:latest"));
-	}
-
-}
